+++ src/main/java/com/gabriel/desafiopicpay/domain/dto/request/TransactionRequest.java
@@ -3,7 +3,9 @@ package com.gabriel.desafiopicpay.domain.dto.request;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
-public record TransactionRequest(@NotNull @Positive Integer value,
+import java.math.BigDecimal;
+
+public record TransactionRequest(@NotNull @Positive BigDecimal value,
                                  @NotNull Integer payer,
                                  @NotNull Integer payee) {
