+++ src/main/java/com/gabriel/desafiopicpay/domain/dto/request/UserRequest.java
@@ -5,11 +5,13 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
+
 public record UserRequest(@NotBlank String name,
                           @NotBlank String document,
                           @NotBlank String email,
                           @NotBlank String password,
                           @NotNull UserType userType,
-                          @PositiveOrZero @NotNull Integer balance) {
+                          @PositiveOrZero @NotNull BigDecimal balance) {
 
 }
