+++ src/main/java/com/gabriel/desafiopicpay/domain/dto/response/TransactionResponse.java
@@ -2,12 +2,13 @@ package com.gabriel.desafiopicpay.domain.dto.response;
 
 import com.gabriel.desafiopicpay.domain.model.enums.StatusTransaction;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record TransactionResponse(UUID id,
                                   String payer,
                                   String payee,
-                                  Integer value,
+                                  BigDecimal value,
                                   StatusTransaction statusTransaction) {
 
 }
